Extract useFakeChatAdapters
